Rename to quickFixesProviders
